adds HMR to map page
